consensus (refactor): address remaining comments on PR

+++ consensus/e2e_tests/pacemaker_test.go
@@ -141,7 +141,7 @@ func TestPacemakerCatchupSameStepDifferentRounds(t *testing.T) {
 
 	// Starting point
 	testHeight := uint64(3)
-	testStep := uint64(consensus.NewRound)
+	testStep := uint8(consensus.NewRound)
 
 	// UnitTestNet configs
 	paceMakerTimeoutMsec := uint64(500) // Set a small pacemaker timeout
